requirements.txt
playwright
pytest
pytest-playwright
openpyxl
python-dotenv
pytest-timeout
streamlit
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
